Refactor Word2Vec config and add example
